reorg activation
